chore: Reorganize frontend directory
