+++ frontend/src/components/AvailableRoomList/AvailableRoomList.test.tsx
@@ -4,7 +4,7 @@ import ReactDOM, { unmountComponentAtNode } from 'react-dom';
 import { DateTime } from 'luxon';
 import { fireEvent, render, screen, waitFor } from '@testing-library/react';
 import AvailableRoomList from './AvailableRoomList';
-import { makeBooking } from '../services/bookingService';
+import { makeBooking } from '../../services/bookingService';
 
 const fakeRooms = [
     {
