+++ frontend/src/components/BookingDrawer/BookingDrawer.tsx
@@ -4,9 +4,9 @@ import AddIcon from '@mui/icons-material/Add';
 import RemoveIcon from '@mui/icons-material/Remove';
 import { DateTime } from 'luxon';
 
-import SwipeableEdgeDrawer, { DrawerContent } from './SwipeableEdgeDrawer';
-import { Room } from '../types';
-import { getTimeLeft, getTimeLeftMinutes2 } from './util/TimeLeft';
+import SwipeableEdgeDrawer, { DrawerContent } from '../BookingView/SwipeableEdgeDrawer';
+import { Room } from '../../types';
+import { getTimeLeft, getTimeLeftMinutes2 } from '../util/TimeLeft';
 
 const MIN_DURATION = 15;
