+++ frontend/src/components/BookingView/DurationPicker.tsx
