+++ frontend/src/components/BookingView/SwipeableEdgeDrawer.tsx
