+++ frontend/src/components/CurrentBooking/CurrentBooking.test.tsx
@@ -4,7 +4,7 @@ import { render, screen, waitFor } from '@testing-library/react';
 import CurrentBooking from './CurrentBooking';
 import userEvent from '@testing-library/user-event';
 import { unmountComponentAtNode } from 'react-dom';
-import { updateBooking, endBooking } from '../services/bookingService';
+import { updateBooking, endBooking } from '../../services/bookingService';
 
 jest.mock('../hooks/useCreateNotification', () => () => {
     return {
