+++ frontend/src/components/UserDrawer/UserDrawer.tsx
@@ -1,9 +1,9 @@
 import { useHistory } from 'react-router-dom';
 import { Visibility } from '@mui/icons-material';
-import SwipeableEdgeDrawer, { DrawerContent } from './SwipeableEdgeDrawer';
-import { DrawerButtonSecondary } from './BookingDrawer';
-import { logout } from '../services/authService';
-import useCreateNotification from '../hooks/useCreateNotification';
+import SwipeableEdgeDrawer, { DrawerContent } from '../BookingView/SwipeableEdgeDrawer';
+import { DrawerButtonSecondary } from '../BookingDrawer/BookingDrawer';
+import { logout } from '../../services/authService';
+import useCreateNotification from '../../hooks/useCreateNotification';
 import { Box } from '@mui/material';
 
 type userSettingsProps = {
